pyproject.toml
[tool.poetry]
name = "drivereader"
version = "0.2.0"
description = "Package reads files in drive and categorizes the data for academic purposes."
authors = ["TheParadox543"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.80.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
openpyxl = "^3.1.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
jupyter = "^1.0.0"
myst-nb = {version = "^0.17.1", python = "^3.11"}
sphinx-autoapi = "^2.0.1"
sphinx-rtd-theme = "^1.2.0"
python-semantic-release = "^7.33.2"
isort = "^5.12.0"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
